SAMSON-642 Fix broken compilation in OSX
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@2577 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/samsonMemCheck/main_SamsonMemCheck.cpp
@@ -71,7 +71,7 @@ void sysctl_value(char *param_name, long int *param_value)
 #ifdef OSX
 void sysctl_value(char *param_name, long int *param_value)
 {
-    long len = sizeof(&(param_value));
+    size_t len = sizeof(&(param_value));
 
     if (sysctlbyname(param_name, param_value, &len, 0, 0) )
     {
